Fix compilation

+++ src/fheroes2/editor/editor_map_specs_window.cpp
@@ -1349,7 +1349,7 @@ namespace
 
                         // Allow to select player only if it is not already selected.
                         if ( le.MouseClickLeft( checkboxRect ) && ( ( _alliances[i] & color ) == 0 )
-                             && ( ( totalPlayerCont + 1 - Color::Count( _alliances[i] ) ) > _alliancesCheckboxes.size() ) ) {
+                             && ( ( totalPlayerCont + 1 - Color::Count( _alliances[i] ) ) > static_cast<int>( _alliancesCheckboxes.size() ) ) ) {
                             for ( size_t ii = 0; ii < _alliancesCheckboxes.size(); ++ii ) {
                                 if ( _alliancesCheckboxes[ii][j]->toggle() ) {
                                     _alliances[ii] |= color;
